NSFW-Backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
transformers
torch
pillow
nudenet
numpy
tf-keras